Get not imported custom tokens from the blockchain
