+++ src/views/activity/activity.view.tsx
@@ -9,14 +9,13 @@ import { useProvidersContext } from "src/contexts/providers.context";
 import { useEnvContext } from "src/contexts/env.context";
 import { useErrorContext } from "src/contexts/error.context";
 import { parseError } from "src/adapters/error";
-import { getBridges } from "src/adapters/bridge-api";
 import { isMetamaskUserRejectedRequestError } from "src/utils/types";
 import { AUTO_REFRESH_RATE } from "src/constants";
 import { Bridge } from "src/domain";
 
 const Activity: FC = () => {
   const env = useEnvContext();
-  const { claim } = useBridgeContext();
+  const { getBridges, claim } = useBridgeContext();
   const { account, connectedProvider } = useProvidersContext();
   const { notifyError } = useErrorContext();
   const [bridgeList, setBridgeList] = useState<Bridge[]>([]);
@@ -69,7 +68,7 @@ const Activity: FC = () => {
         clearInterval(intervalId);
       };
     }
-  }, [account, env, notifyError]);
+  }, [account, env, getBridges, notifyError]);
 
   useEffect(() => {
     setWrongNetworkBridges([]);
